fix serialization type errors
